split methods into their own file
